Revert "Fix log interception to respect TF_LOG  (#1537)"
This reverts commit 48a49522ddcac0d86608d53670b836d8ab2eb5f0.
